pyproject.toml
[project]
name = "agent-framework-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agent-framework",
    "python-dotenv>=1.2.1",
]


[build-system]
requires = ["uv_build>=0.9.3,<0.10.0"]
build-backend = "uv_build"

